_scripts/requirements.txt
numbers-parser
python-frontmatter
pyyaml
requests>=2.31.0
python-dotenv>=1.0.0
responses>=0.24.0
google-api-python-client>=2.0.0
rapidfuzz>=3.0.0
